# packages/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExercisePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          }
        },
        "required": [
          "items",
          "next"
        ],
        "title": "ExercisePage",
        "type": "object"
      },
      "ExerciseResponse": {
        "description": "A global exercise carries a translation key and no name; the user's own carries a name and\nno key (INV-27). `is_own` says which.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_max_reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Max Reps"
          },
          "default_min_reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Min Reps"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "forked_from_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forked From Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_own": {
            "title": "Is Own",
            "type": "boolean"
          },
          "is_unilateral": {
            "title": "Is Unilateral",
            "type": "boolean"
          },
          "load_increment_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Increment Kg"
          },
          "modality": {
            "enum": [
              "barbell",
              "dumbbell",
              "machine",
              "cable",
              "bodyweight",
              "band",
              "other"
            ],
            "title": "Modality",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "name_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Key"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "primary_muscle_id": {
            "title": "Primary Muscle Id",
            "type": "integer"
          },
          "secondary_muscle_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Secondary Muscle Ids",
            "type": "array"
          },
          "tracking": {
            "enum": [
              "weight_reps",
              "reps_only",
              "duration",
              "distance_duration"
            ],
            "title": "Tracking",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uses_bodyweight": {
            "title": "Uses Bodyweight",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "is_own",
          "forked_from_id",
          "name",
          "name_key",
          "modality",
          "primary_muscle_id",
          "secondary_muscle_ids",
          "is_unilateral",
          "tracking",
          "load_increment_kg",
          "uses_bodyweight",
          "default_min_reps",
          "default_max_reps",
          "notes",
          "created_at",
          "updated_at",
          "deleted_at"
        ],
        "title": "ExerciseResponse",
        "type": "object"
      },
      "ExerciseWrite": {
        "additionalProperties": false,
        "description": "One of the user's own exercises. The name is stored exactly as typed and never translated\n(INV-27); a fork names the exercise it came from (stage 4).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_max_reps": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Max Reps"
          },
          "default_min_reps": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Min Reps"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "forked_from_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forked From Id"
          },
          "is_unilateral": {
            "default": false,
            "title": "Is Unilateral",
            "type": "boolean"
          },
          "load_increment_kg": {
            "anyOf": [
              {
                "maximum": 9999.999999,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Increment Kg"
          },
          "modality": {
            "enum": [
              "barbell",
              "dumbbell",
              "machine",
              "cable",
              "bodyweight",
              "band",
              "other"
            ],
            "title": "Modality",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "primary_muscle_id": {
            "maximum": 32767.0,
            "minimum": 1.0,
            "title": "Primary Muscle Id",
            "type": "integer"
          },
          "secondary_muscle_ids": {
            "items": {
              "maximum": 32767.0,
              "minimum": 1.0,
              "type": "integer"
            },
            "maxItems": 20,
            "title": "Secondary Muscle Ids",
            "type": "array"
          },
          "tracking": {
            "default": "weight_reps",
            "enum": [
              "weight_reps",
              "reps_only",
              "duration",
              "distance_duration"
            ],
            "title": "Tracking",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uses_bodyweight": {
            "default": false,
            "title": "Uses Bodyweight",
            "type": "boolean"
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "name",
          "modality",
          "primary_muscle_id"
        ],
        "title": "ExerciseWrite",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkoutPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkoutSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          }
        },
        "required": [
          "items",
          "next"
        ],
        "title": "WorkoutPage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CyberAthlete API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exercises": {
      "get": {
        "description": "The global catalog and the user's own exercises, by id, a page at a time.",
        "operationId": "list_exercises_api_v1_exercises_get",
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExercisePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Exercises",
        "tags": [
          "exercises"
        ]
      }
    },
    "/api/v1/exercises/{exercise_id}": {
      "get": {
        "description": "A global or one of the user's own. Someone else's answers 404, like an absent one (04 \u00a74).",
        "operationId": "get_exercise_api_v1_exercises__exercise_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "exercise_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exercise Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Exercise",
        "tags": [
          "exercises"
        ]
      },
      "put": {
        "description": "Creates or updates one of the user's own exercises (decision 3). The copy with the newer\n`updated_at` wins; a global's id, like somebody else's, is `409 id_unavailable`.",
        "operationId": "put_exercise_api_v1_exercises__exercise_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "exercise_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exercise Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            },
            "description": "Created"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Exercise",
        "tags": [
          "exercises"
        ]
      }
    },
...
    "/api/v1/workouts": {
      "get": {
        "description": "The user's workouts, without their exercises, newest start first, a page at a time.",
        "operationId": "list_workouts_api_v1_workouts_get",
        "parameters": [
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workouts",
        "tags": [
          "workouts"
        ]
      }
    },
    "/api/v1/workouts/{workout_id}": {
      "get": {
        "description": "The workout, every exercise entry and every set, archived ones included. 404 for a workout\nthat does not exist and for one that is somebody else's alike (04 \u00a74).",
        "operationId": "get_workout_api_v1_workouts__workout_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workout Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Workout",
        "tags": [
          "workouts"
        ]
      },
      "put": {
        "description": "Applies the document row by row \u2014 each row's newer `updated_at` wins, a row left out is\nkept \u2014 and answers with the workout as stored (decision 1). Idempotent by the client's ids\n(INV-16): resending changes nothing. A completed set without its mode's measure is\n`422 set_missing_<field>` (decision 2).",
        "operationId": "put_workout_api_v1_workouts__workout_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "workout_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workout Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            },
            "description": "Created"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Workout",
        "tags": [
          "workouts"
        ]
      }
    },
...
  }
}